# packages/api-client/src/generated/model/audioUploadUrlDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * ScrAI — AI Scribe Notes API
 * Create and view AI-generated clinical notes associated with patients.
 * OpenAPI spec version: 0.1.0
 */

export interface AudioUploadUrlDto {
  /** S3 object key to pass back when creating the note */
  key: string;
  /** Presigned PUT URL the browser uploads the file to */
  url: string;
}