# packages/api-client/src/generated/model/createAudioNoteDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * ScrAI — AI Scribe Notes API
 * Create and view AI-generated clinical notes associated with patients.
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAudioNoteDto {
  patientId: string;
  /** S3 object key returned from /notes/audio/upload-url */
  audioKey: string;
  /** Original filename, stored for display */
  audioFilename: string;
  title?: string;
  /** Generate an AI SOAP summary from the transcription */
  summarize?: boolean;
}