# packages/api-client/src/generated/model/createAudioUploadUrlDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * ScrAI — AI Scribe Notes API
 * Create and view AI-generated clinical notes associated with patients.
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAudioUploadUrlDto {
  /** Original filename (used to build the object key) */
  filename: string;
  /** MIME type; the PUT must send this same Content-Type */
  contentType: string;
}